+++ holiday-manager-core/src/main/java/holiday/manager/application/service/holiday/application/HolidayApplicationService.java
@@ -114,52 +114,4 @@ public class HolidayApplicationService {
 		return application;
 
 	}
-
-	public List<HolidayApplication> findByAplicantId(UserId applicantId) {
-
-		List<HolidayApplicationEntity> holidayApplicationEntities = queryRepository.findByAplicantId(applicantId.getValue());
-
-		return convert(holidayApplicationEntities);
-	}
-
-	public List<HolidayApplication> findByStatus(HolidayApplicationStatus status) {
-
-		List<HolidayApplicationEntity> holidayApplicationEntities = queryRepository.findByStatus(status.name());
-
-		return convert(holidayApplicationEntities);
-	}
-
-	public List<HolidayApplication> findByAplicantIdAndStatus(UserId applicantId, HolidayApplicationStatus status) {
-
-		List<HolidayApplicationEntity> holidayApplicationEntities = queryRepository.findByAplicantIdAndStatus(applicantId.getValue(), status.name());
-
-		return convert(holidayApplicationEntities);
-	}
-
-	private List<HolidayApplication> convert(List<HolidayApplicationEntity> holidayApplicationEntities) {
-
-		List<HolidayApplication> holidayApplications = new ArrayList<HolidayApplication>();
-
-		for (HolidayApplicationEntity entity : holidayApplicationEntities) {
-			addHolidayApplication(holidayApplications, entity);
-		}
-
-		return holidayApplications;
-	}
-
-	private void addHolidayApplication(List<HolidayApplication> holidayApplications, HolidayApplicationEntity entity) {
-
-		HolidayApplicationId holidayApplicationId = new HolidayApplicationId(entity.getId());
-		KindOfHoliday kindOfHoliday = KindOfHoliday.valueOf(entity.getKind());
-		HolidayType holidayType = HolidayType.valueOf(entity.getType());
-		Date date = entity.getDate();
-		HolidayApplicationStatus holidayApplicationStatus = HolidayApplicationStatus.valueOf(entity.getStatus());
-		UserId applicant = new UserId(entity.getAplicantId());
-		UserId approverId = new UserId(entity.getApproverId());
-
-		HolidayApplication holidayApplication = new HolidayApplication(holidayApplicationId, kindOfHoliday, holidayType,
-				date, holidayApplicationStatus, applicant, approverId);
-
-		holidayApplications.add(holidayApplication);
-	}
 }
